Archive/requirements_llm.txt
langchain-community==0.0.13
langchain==0.1.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
cloudinary==1.41.0
python-docx==1.1.0
markdown==3.5.2
watchdog==3.0.0
huggingface-hub==0.20.3
sentence-transformers==2.2.2
transformers==4.36.2
numpy==1.24.3
tqdm==4.66.1
python-magic==0.4.27
torch==2.1.0
accelerate==0.25.0
bitsandbytes==0.41.1 